+++ src/cmd/validate-values.ts
@@ -10,7 +10,7 @@ import { Arguments, helmOptions } from '../common/yargs-opts'
 const cmdName = getFilename(__filename)
 const debug: OtomiDebugger = terminal(cmdName)
 
-const internalPaths: string[] = ['apps', 'k8s', 'services', 'sops', 'teamConfig.services']
+const internalPaths: string[] = ['apps', 'k8s', 'services', 'teamConfig.services']
 
 // TODO: Accept json path to validate - on empty, validate all
 export const validateValues = async (): Promise<void> => {
@@ -32,7 +32,7 @@ export const validateValues = async (): Promise<void> => {
   }
 
   debug.info('Loading values-schema.yaml')
-  const valuesSchema = loadYaml(`${rootDir}/values-schema.yaml`) as Record<string, unknown>
+  const valuesSchema = loadYaml(`${rootDir}/values-schema.yaml`) as Record<string, any>
   debug.debug('Initializing Ajv')
   const ajv = new Ajv({ allErrors: true, strict: false, strictTypes: false, verbose: true })
   debug.debug('Compiling Ajv validation')