+++ src/common/values.ts
@@ -10,7 +10,7 @@ const objectToYaml = (obj: Record<string, any>): string => {
   return isEmpty(obj) ? '' : dump(obj, { indent: 4 })
 }
 
-const removeBlankAttributes = (obj: Record<string, unknown>): Record<string, unknown> => {
+const removeBlankAttributes = (obj: Record<string, any>): Record<string, any> => {
   const options: CleanOptions = {
     emptyArrays: false,
     emptyObjects: true,
@@ -65,7 +65,7 @@ export const writeValues = async (values: Record<string, any>, overwrite = true)
   d.debug('secrets: ', JSON.stringify(secrets, null, 2))
   // removing secrets
   const plainValues = removeBlankAttributes(omit(values, secretsJsonPath)) as any
-  const fieldsToOmit = ['cluster', 'policies', 'teamConfig', 'charts']
+  const fieldsToOmit = ['cluster', 'policies', 'teamConfig', 'charts', '_derived']
   const secretSettings = omit(secrets, fieldsToOmit)
   const settings = omit(plainValues, fieldsToOmit)
 
@@ -80,7 +80,7 @@ export const writeValues = async (values: Record<string, any>, overwrite = true)
   if (plainValues.policies)
     promises.push(writeValuesToFile(`${env.ENV_DIR}/env/policies.yaml`, { policies: plainValues.policies }, overwrite))
 
-  const plainChartPromises = Object.keys(plainValues.charts || {}).map((chart) => {
+  const plainChartPromises = Object.keys((plainValues.charts || {}) as Record<string, any>).map((chart) => {
     const valueObject = {
       charts: {
         [chart]: plainValues.charts[chart],
@@ -88,10 +88,10 @@ export const writeValues = async (values: Record<string, any>, overwrite = true)
     }
     return writeValuesToFile(`${env.ENV_DIR}/env/charts/${chart}.yaml`, valueObject, overwrite)
   })
-  const secretChartPromises = Object.keys((secrets.charts || {}) as Record<string, unknown>).map((chart) => {
+  const secretChartPromises = Object.keys((secrets.charts || {}) as Record<string, any>).map((chart) => {
     const valueObject = {
       charts: {
-        [chart]: values.charts[chart],
+        [chart]: secrets.charts[chart],
       },
     }
     return writeValuesToFile(`${env.ENV_DIR}/env/charts/secrets.${chart}.yaml`, valueObject, overwrite)